change all Integers to Longs

+++ src/main/java/traveller/dtos/DriverDTO.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public class DriverDTO {
 
-    private Integer id;
+    private Long id;
     @NotNull @NotBlank @Size(min = 3)
     private String firstName;
     @NotNull @NotBlank @Size(min = 3)
@@ -20,10 +20,11 @@ public class DriverDTO {
     private String email;
     private List<Tour> tours;
 
-    public Integer getId() {
+    public Long getId() {
+
         return id;
     }
-    public void setId(Integer id) {
+    public void setId(Long id) {
 
         this.id = id;
     }
